course_enrollment/requirements.txt
Django==5.0.14
djangorestframework
django-cors-headers
